Fix architecture bit width-dependent tests
555_666_7777 is 33 bits wide which overflows `usize` on 32-bit targets.

+++ tests/rustc-deriving-hash.rs
@@ -19,9 +19,9 @@ use std::collections::hash_map::DefaultHasher;
 #[derive(Derivative)]
 #[derivative(Hash)]
 struct Person {
-    id: usize,
+    id: u16,
     name: String,
-    phone: usize,
+    phone: u64,
 }
 
 // test for hygiene name collisions
